new commit

+++ src/main/java/br/ufpb/dcx/dsc/figurinhas/services/AlbumService.java
@@ -1,26 +1,28 @@
 package br.ufpb.dcx.dsc.figurinhas.services;
 
+import br.ufpb.dcx.dsc.figurinhas.exceptions.AlbumNotFoundException;
+import br.ufpb.dcx.dsc.figurinhas.exceptions.UserNotFoundException;
 import br.ufpb.dcx.dsc.figurinhas.models.Album;
 import br.ufpb.dcx.dsc.figurinhas.models.User;
 import br.ufpb.dcx.dsc.figurinhas.repository.AlbumRepository;
 import br.ufpb.dcx.dsc.figurinhas.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
-import java.util.*;
+import java.util.List;
 
 @Service
 public class AlbumService {
-    private AlbumRepository albumRepository;
-    private UserRepository userRepository;
+    private final AlbumRepository albumRepository;
+    private final UserRepository userRepository;
 
-    public AlbumService(AlbumRepository albumRepository, UserRepository userRepository)
-    {
+    public AlbumService(AlbumRepository albumRepository, UserRepository userRepository) {
         this.albumRepository = albumRepository;
         this.userRepository = userRepository;
     }
 
     public Album getAlbum(Long id){
-        return albumRepository.getReferenceById(id);
+        return albumRepository.findById(id)
+                .orElseThrow(() -> new AlbumNotFoundException(id));
     }
 
     public List<Album> listAlbuns() {
@@ -28,25 +30,26 @@ public class AlbumService {
     }
 
     public Album saveAlbum(Album a, Long userId) {
-        Optional<User> userOpt = userRepository.findById(userId);
-        if(userOpt.isPresent()){
-            a.setUser(userOpt.get());
-            return albumRepository.save(a);
-        }
-        return null;
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new UserNotFoundException(userId));
+        a.setUser(user);
+        return albumRepository.save(a);
     }
 
     public void deleteAlbum(Long id) {
-        albumRepository.deleteById(id);
+        if(albumRepository.existsById(id)) {
+            albumRepository.deleteById(id);
+        } else {
+            throw new AlbumNotFoundException(id);
+        }
     }
 
     public Album updateAlbum(Long id, Album f) {
-        Optional<Album> figOpt = albumRepository.findById(id);
-        if(figOpt.isPresent()){
-            Album toUpdate = figOpt.get();
-            toUpdate.setNome(f.getNome());
-            return toUpdate;
-        }
-        return null;
+        return albumRepository.findById(id)
+                .map(album -> {
+                    album.setNome(f.getNome());
+                    return albumRepository.save(album);
+                })
+                .orElseThrow(() -> new AlbumNotFoundException(id));
     }
 }
