+++ src/main/java/br/ufpb/dcx/dsc/figurinhas/services/FigurinhaService.java
@@ -1,22 +1,23 @@
 package br.ufpb.dcx.dsc.figurinhas.services;
 
+import br.ufpb.dcx.dsc.figurinhas.exceptions.FigurinhaNotFoundException;
 import br.ufpb.dcx.dsc.figurinhas.models.Figurinha;
 import br.ufpb.dcx.dsc.figurinhas.repository.FigurinhaRepository;
 import org.springframework.stereotype.Service;
 
-import java.util.*;
+import java.util.List;
 
 @Service
 public class FigurinhaService {
-    private final ArrayList<Figurinha> figurinhaList= new ArrayList<>();
-    private FigurinhaRepository figurinhaRepository;
+    private final FigurinhaRepository figurinhaRepository;
 
     public FigurinhaService(FigurinhaRepository figurinhaRepository){
         this.figurinhaRepository = figurinhaRepository;
     }
 
     public Figurinha getFigurinha(Long id){
-        return figurinhaRepository.getReferenceById(id);
+        return figurinhaRepository.findById(id)
+                .orElseThrow(() -> new FigurinhaNotFoundException(id));
     }
 
     public List<Figurinha> listFigurinhas() {
@@ -28,17 +29,20 @@ public class FigurinhaService {
     }
 
     public void deleteFigurinha(Long id) {
-        figurinhaRepository.deleteById(id);
+        if(figurinhaRepository.existsById(id)){
+            figurinhaRepository.deleteById(id);
+        } else {
+            throw new FigurinhaNotFoundException(id);
+        }
     }
 
     public Figurinha updateFigurinha(Long id, Figurinha f) {
-        Optional<Figurinha> figOpt = figurinhaRepository.findById(id);
-        if(figOpt.isPresent()){
-            Figurinha toUpdate = figOpt.get();
-            toUpdate.setSelecao(f.getSelecao());
-            toUpdate.setNome(f.getNome());
-            return toUpdate;
-        }
-        return null;
+        return figurinhaRepository.findById(id)
+                .map(toUpdate -> {
+                    toUpdate.setSelecao(f.getSelecao());
+                    toUpdate.setNome(f.getNome());
+                    return figurinhaRepository.save(toUpdate);
+                })
+                .orElseThrow(() -> new FigurinhaNotFoundException(id));
     }
 }
